Fix. SFW. Reduction to a number.

+++ cleantalk.antispam/lib/Cleantalk/Common/Firewall/FirewallUpdater.php
@@ -97,7 +97,7 @@ class FirewallUpdater
         if(
             Get::get('spbc_remote_call_action') == 'sfw_update__write_base' &&
             Get::get('firewall_updating_id') &&
-            Get::get('firewall_updating_id') !== $fw_stats['firewall_updating_id'] &&
+            (int)Get::get('firewall_updating_id') !== (int)$fw_stats['firewall_updating_id'] &&
             time() - $fw_stats['firewall_updating_last_start'] < 8600
         ) {
             return array( 'error' => 'FIREWALL_IS_UPDATING' );
